+++ src/lib/chat.js
@@ -11,7 +11,7 @@ import chatCss from '../css/chat.css';
 let concat = function(...args) {
   return args.reduce(function(result, item) {
     // If array-like
-    if(item.length && !Array.isArray(item)) {
+    if(item && item.length && !Array.isArray(item)) {
       item = Array.prototype.slice.call(item);
     }
     
@@ -89,8 +89,12 @@ let gitterSvgSprites = `
       <symbol id="gitter-shape-external-link" viewBox="0 0 1792 1792">
         <path d="M1408 928v320q0 119-84.5 203.5t-203.5 84.5h-832q-119 0-203.5-84.5t-84.5-203.5v-832q0-119 84.5-203.5t203.5-84.5h704q14 0 23 9t9 23v64q0 14-9 23t-23 9h-704q-66 0-113 47t-47 113v832q0 66 47 113t113 47h832q66 0 113-47t47-113v-320q0-14 9-23t23-9h64q14 0 23 9t9 23zm384-864v512q0 26-19 45t-45 19-45-19l-176-176-652 652q-10 10-23 10t-23-10l-114-114q-10-10-10-23t10-23l652-652-176-176q-19-19-19-45t19-45 45-19h512q26 0 45 19t19 45z"/>
       </symbol>
+      <symbol id="gitter-shape-spinner" viewBox="0 0 1792 1792">
+        <path d="M526 1394q0 53-37.5 90.5t-90.5 37.5q-52 0-90-38t-38-90q0-53 37.5-90.5t90.5-37.5 90.5 37.5 37.5 90.5zm498 206q0 53-37.5 90.5t-90.5 37.5-90.5-37.5-37.5-90.5 37.5-90.5 90.5-37.5 90.5 37.5 37.5 90.5zm-704-704q0 53-37.5 90.5t-90.5 37.5-90.5-37.5-37.5-90.5 37.5-90.5 90.5-37.5 90.5 37.5 37.5 90.5zm1202 498q0 52-38 90t-90 38q-53 0-90.5-37.5t-37.5-90.5 37.5-90.5 90.5-37.5 90.5 37.5 37.5 90.5zm-964-996q0 66-47 113t-113 47-113-47-47-113 47-113 113-47 113 47 47 113zm1170 498q0 53-37.5 90.5t-90.5 37.5-90.5-37.5-37.5-90.5 37.5-90.5 90.5-37.5 90.5 37.5 37.5 90.5zm-640-704q0 80-56 136t-136 56-136-56-56-136 56-136 136-56 136 56 56 136zm530 206q0 93-66 158.5t-158 65.5q-93 0-158.5-65.5t-65.5-158.5q0-92 65.5-158t158.5-66q92 0 158 66t66 158z"/>
+      </symbol>
     </defs>
-  </svg>`;
+  </svg>
+`;
 
 let embedGitterSvgSprites = function() {
   let elementStore = new ElementStore();
@@ -109,13 +113,7 @@ let embedGitterSvgSprites = function() {
 let embedGitterChat = function(opts) {
   let elementStore = new ElementStore();
 
-  let containers = coerceIntoElementsArray(opts.container || (() => {
-    let container = elementStore.createElement('aside');
-    container.classList.add('gitter-chat-embed');
-    document.body.appendChild(container);
-
-    return container;
-  })());
+  let containers = opts.container;
 
   containers.forEach((container) => {
     let containerOpts = getDataOptionsFromElement(opts, container);
@@ -134,10 +132,7 @@ let embedGitterChat = function(opts) {
 
   });
 
-  return {
-    containers,
-    elementStore
-  };
+  return elementStore;
 };
 
 
@@ -157,7 +152,7 @@ let defaults = {
   activation: null,
   // Whether to preload the gitter chat iframe.
   // We preload the chat so there isn't any jank when the chat opens
-  preload: true,
+  preload: false,
 
   // Whether to embed a `<style>` tag with some pre-made CSS
   useStyles: true,
@@ -169,6 +164,7 @@ let defaults = {
 
   //showLeftMenu: false
 };
+// Make the defaults a little more immutable
 Object.keys(defaults).forEach((key) => {
   Object.defineProperty(defaults, key, {
     value: defaults[key],
@@ -177,24 +173,37 @@ Object.keys(defaults).forEach((key) => {
   });
 });
 
+
 // Keep some stuff behind symbols so people "can't" access the private data
 const DEFAULTS = Symbol();
 const OPTIONS = Symbol();
 const ELEMENTSTORE = Symbol();
 const INIT = Symbol();
-const CONTAINERS = Symbol();
+const ISEMBEDDED = Symbol();
 const EMBEDCHATONCE = Symbol();
-class chatEmbed {
+const TOGGLECONTAINERS = Symbol();
 
+class chatEmbed {
   constructor(options = {}) {
     this[ELEMENTSTORE] = new ElementStore();
     
-    this[DEFAULTS] = defaults;
-
+    this[DEFAULTS] = objectAssign({}, defaults);
 
     // Coerce into array of dom elements on what they pass in
     if(options.container) {
       options.container = coerceIntoElementsArray(options.container);
+    }
+    // Otherwise create our own default container
+    else {
+      this[DEFAULTS].container = coerceIntoElementsArray((() => {
+        let container = this[ELEMENTSTORE].createElement('aside');
+        container.classList.add('gitter-chat-embed');
+        // Start out collapsed
+        container.classList.add('is-collapsed');
+        document.body.appendChild(container);
+
+        return container;
+      })());
     }
 
     this[OPTIONS] = objectAssign({}, this[DEFAULTS], options);
@@ -210,10 +219,21 @@ class chatEmbed {
       this[ELEMENTSTORE] = this[ELEMENTSTORE].concat(embedGitterSvgSprites());
     }
 
+    let containers = opts.container;
+    containers.forEach((container) => {
+      let loadingIndicatorElement = this[ELEMENTSTORE].createElement('div');
+      loadingIndicatorElement.classList.add('gitter-chat-embed-loading-wrapper');
+      loadingIndicatorElement.innerHTML = `
+        <svg class=" gitter-chat-embed-loading-indicator gitter-icon"><use xlink:href="#gitter-shape-spinner"></use></svg>
+      `;
+
+      // Prepend
+      container.insertBefore(loadingIndicatorElement, container.firstChild);
+    });
+
     if(opts.preload) {
       this.toggleChat(false);
     }
-
 
     if(opts.showChatByDefault) {
       this.toggleChat(true);
@@ -239,7 +259,7 @@ class chatEmbed {
             e.preventDefault();
           });
 
-          this[CONTAINERS].forEach((container) => {
+          opts.container.forEach((container) => {
             container.on('gitter-chat-toggle', (e) => {
               let isChatOpen = e.detail.state;
               // Toggle the visibiltiy of the activation element
@@ -254,16 +274,18 @@ class chatEmbed {
   }
 
   [EMBEDCHATONCE]() {
-    if(!this[CONTAINERS]) {
+    if(!this[ISEMBEDDED]) {
+      let opts = this[OPTIONS];
+
       let embedResult = embedGitterChat(this[OPTIONS]);
-      this[CONTAINERS] = embedResult.containers;
-      this[ELEMENTSTORE] = this[ELEMENTSTORE].concat(embedResult.elementStore);
+      this[ELEMENTSTORE] = this[ELEMENTSTORE].concat(embedResult);
 
-
-      this[CONTAINERS].forEach((container) => {
+      let containers = opts.container;
+      containers.forEach((container) => {
         let actionBar = this[ELEMENTSTORE].createElement('div');
         actionBar.classList.add('gitter-chat-embed-action-bar');
 
+        // Prepend
         container.insertBefore(actionBar, container.firstChild);
 
         let collapseActionElement = this[ELEMENTSTORE].createElement('button');
@@ -299,19 +321,20 @@ class chatEmbed {
         
 
       });
-
     }
+
+    this[ISEMBEDDED] = true;
   }
 
+  [TOGGLECONTAINERS](state) {
+    let opts = this[OPTIONS];
 
-  // Public API
-  toggleChat(state) {
-    this[EMBEDCHATONCE]();
-    if(!this[CONTAINERS]) {
+    if(!opts.container) {
       console.warn('Gitter Sidecar: No chat embed elements to toggle visibility on');
     }
 
-    coerceIntoElementsArray(this[CONTAINERS]).forEach(function(container) {
+    let containers = opts.container;
+    containers.forEach((container) => {
       container.classList.toggle('is-collapsed', !state);
 
       let event = new CustomEvent('gitter-chat-toggle', {
@@ -321,6 +344,38 @@ class chatEmbed {
       });
       container.dispatchEvent(event);
     });
+  }
+
+
+  // Public API
+  toggleChat(state) {
+    let opts = this[OPTIONS];
+
+    // We delay the embed to make sure the animation can go jank free
+    // if it isn't already embedded
+    if(state && !this[ISEMBEDDED]) {
+      let containers = opts.container;
+      // Start the loading spinner
+      containers.forEach((container) => {
+        container.classList.add('is-loading');
+      });
+
+      setTimeout(() => {
+        this[EMBEDCHATONCE]();
+        this[TOGGLECONTAINERS](state);
+
+        // Remove the loading spinner
+        containers.forEach((container) => {
+          container.classList.remove('is-loading');
+        });
+      }, 300/* TODO change to transition/animation end, see for robust transition/animation end code: https://github.com/MadLittleMods/jquery-carouselss */);
+    }
+    // But we still want people to embed no matter what state :)
+    // For example `options.preload`, should load the chat but not show it
+    else {
+      this[EMBEDCHATONCE]();
+      this[TOGGLECONTAINERS](state);
+    }
   }
 
   destroy() {